widgets: Defer repaint on screen/dpr change to expose/update request
When a window is moved from one screen to another, or the DPR of the
screen changes, the QPA layer may emit a screen or DPR change event
for the QWindow.

This event, just like resize events, is not a request to paint the
window. It's information to the window that may or may not result
in the window requiring a repaint. If so, the window should request
an update explicitly.

If the platform itself determines that a screen or DPR change will
require a repaint, it will deliver an expose event explicitly.

Both the expose event and update request will be synchronized with
the platform's drawing cycle, as opposed to drawing from a callback
such as resize events or screen/DPR change events.

Task-number: QTBUG-114873
Change-Id: I2fb78f177d84800f43764c36e85b554acce65a00
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: David Edmundson <davidedmundson@kde.org>
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>