AndroidDeployQt: Remove not satisfied dependencies
In 438837ce274fdc1457b66179b25df40f33b23a15 commit, we stopped to copy
all dependent libs that are kept in plugin directory. This change sounds
reasonable, as we do not want to deploy unused libs.

Due to the mentioned change, when dependencies.xml contains *.so
libraries, they are not copied at all now. They are added to localLibs
(later pasted into libs.xml), but are not shipped with the apk.
So at the end we have dependencies in libs.xml on some libraries that
were not delievered.

This commit cleans up the localLibs to not add dependencies to the
libs.xml file as they will not be satisfied

Fixes: QTBUG-129946
Pick-to: 6.8
Change-Id: I7157e2e65cb928adbe9d7c077e50b2ebcac94490
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>