Implement helper function for handling key events in QGroupBox
A GroupBox would for its' checkbox accept a key event if this event
was a part of buttonPressKeys. This would even happen if the QGroupBox
wasn't checkable, which resulted with an unnecessarily consumed
key event.

Implement a helper function QGroupBoxPrivate::shouldHandleKeyEvent()
that can be used to check if a key event should be accepted,
and that takes into consideration if the QGroupBox is checkable.

Fixes: QTBUG-129095
Change-Id: I21d1aa24b1b25b70d5274e2c13f887aefc80dbc5
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>