Enforce sorting of QTZ backend availableTimeZoneIds()
Various pieces of QTZP and QTZ rely (by using std::set_union or
std::set_intersection) on the backends supplying the list of available
IDs in sorted order. As this isn't documented as part of public API
behavior, document it internally and assert it.

Add the missing sorting and uniquification for Android; although the
Java API used does appear to return a sorted list, its specification
[0] says nothing about its ordering, so make sure. In the process,
rename a wantonly long variable so the uniquifying doesn't need a
line-split.

[0] https://docs.oracle.com/javase/8/docs/api/java/util/TimeZone.html#getAvailableIDs--

Document, in the ICU backend, that the relevant API doesn't say what
order it returns; but also make clear where the backend *does* take
care of sorting it and rendering entries unique.

Change-Id: I61b4ecf342044e2303b18889c600da5922aec278
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>