a11y atspi: Send AT-SPI signals with correct signature
The AT-SPI2 XML spec for all signals/events [1] that Qt emits
specifies the DBus arguments signature as "siiva{sv}".

Qt however was incorrectly sending arguments with an "siiv(so)"
signature instead, by sending the QSpiObjectReference as last
argument.

Fix this by sending the events with the proper signature, using an
empty QMap<QString, QVariant> for the properties (that translates
to an "a{sv}" type for DBus).

[1] https://gitlab.gnome.org/GNOME/at-spi2-core/-/blob/33b45aea12331335eaeb7edb45132466114013e9/xml/Event.xml

Fixes: QTBUG-130843
Pick-to: 6.8
Change-Id: If023763626be3d17b091466b17829d62d29fdede
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>