Revert "QTest: add -[no]throwon{fail,skip} command line arguments"
This reverts commit fde57300ab51c7deda168f6a4515dcf7b1340618 and a
portion of commit 627f804d80d76cb729b84dc319382213cd89763c and of commit
e769cf026e328ed7fff660c204ce6e55b80114e3.

The command-line feature and the environment variables are a design
flaw, because there are NO conditions under which a regular user who
isn't the designer of the test can make the correct choice. Either the
tests are designed to work with throwing or they are not. The test must
be run in the mode that the test developer wrote it.

[ChangeLog][QtTest] The command-line options and environment variables
that changed the defaults for QTest::setThrowOn{Fail,Skip}() have been
removed due to a design flaw. To set the default, either place a call to
the required function in the test's initTestCase() function or define
the C++ macro.

Task-number: QTBUG-66320
Change-Id: I0c17899984ebfd70de97fffd51784aa42d41b743
Reviewed-by: Jason McDonald <macadder1@gmail.com>