Ensure system/default locale sets number options suitably when it's C
In the process, introduce inline helpers to avoid repetition of the
ternary that selects the options based on language C or otherwise.

Note that, if the system locale changes under our feet, the default
locale's number options may be left out of sync with it. It's
possible, though kludgy, to work round that; but doing so creates the
inverse problem where the caller has created a copy of the system
locale, set its number options to something other than the default and
set that as default locale. Auto-syncing the default with changes to
the system would then be apt to stomp the carefully set number
options. Given that the system locale (currently, at least) only
changes on program start-up and exit, take the path less kludged.

Change-Id: I23ab38b6e8a6936f85c88ca66245a89612e3c3fc
Reviewed-by: Magdalena Stojek <magdalena.stojek@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>