QThread/Win: replace the adopted thread watcher with a thread_local
The adopted thread watcher thread on Windows was a massive hack: because
Win32's Thread Local Storage support[1] did not support registering a
destructor for the data, we couldn't use the same trick we did on Unix
to run some code on the adopted thread when it was about to terminate.
So instead we launched another thread (which would also be adopted so we
had to protect against being watched) to watch the adopted threads and
post notifications. This consumed more resources and also emitted the
QThread::finished() signal from the watcher thread, not the adopted
one. Plus, this could cause crashes if QtCore was unloaded while the
watcher thread was running.

We can now depend on C++11 thread_locals on Windows, which therefore
allows us to run code shortly before the thread's exit. This commit
copies code from qthread_unix.cpp introduced in commit
4fabde349f16b59f37568da2a4c050c6dd53a34e, not including the two-phase
cleaning up.

This allows us to remove the QThreadData destruction code from
QCoreApplicationData.

[1] https://learn.microsoft.com/en-us/windows/win32/ProcThread/using-thread-local-storage

Change-Id: I09deebfec02a15082a68fffd7f650427eabdee54
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>