Fix blacklisting of parameter test of QColorSpace
When creating blacklists for parametric tests, I've misunderstand the
role of comma in parametric tests descriptor in BLACKLIST, thinking that
this is a separator when in fact it was just a part of parameter name.
This leads to tests not being actually blacklisted for VxWorks.

Fix this by blacklisting each parametric tests separately.

Task-number: QTBUG-130076
Task-number: QTBUG-115777
Change-Id: I7ca48a3f4bace6250aa9c990f5374912d3f3c561
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>