Blacklist QUdpSocket test functions failing on VxWorks
Test functions:
 - writeDatagramToNonExistingPeer
 - multicastTtlOption
 - multicastLoopbackOption
 - readyRead
are failing on VxWorks in its nightly COIN run. We do not know the root
causes of these failures.

We need to run tests on COIN for VxWorks as significant, and fails in
QUdpSocket test functions prevents it, to reduce growing number of
regressions and allow iterative introduction of VxWorks as a supported
platform.

Blacklist QUdpSocket test functions failing on VxWorks.

Task-number: QTBUG-115777
Task-number: QTBUG-130078
Change-Id: I47c18f6d53406d4ecacfd35fb27ef883b561951d
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>