CMake: Store architecture information about all platforms
...in a multi-architecture build on macOS.

Our "architecture" configure test only retrieved and stored information
about the first architecture, even if CMAKE_OSX_ARCHITECTURES contained
multiple values.

Now, we run that configure test once for every entry in
CMAKE_OSX_ARCHITECTURES, extract the information from the test binary,
and store
- a list of architectures,
- a list of subarchitectures per architecture,
- the build ABI per architecture.

For example, for a build of Qt with CMAKE_OSX_ARCHITECTURES=x86_64;arm64
we now set the following cache variables.

For the first architecture:
    TEST_architecture_arch=x86_64
    TEST_subarch_result=cx16;sse3;ssse3;sse4.1

The list of architectures:
    TEST_architecture_architectures=x86_64;arm64

Sub-architectures and ABI per architecture:
    TEST_arch_arm64_abi=arm64-little_endian-lp64
    TEST_arch_arm64_subarch_crc32=1
    TEST_arch_arm64_subarch_crypto=1
    TEST_arch_arm64_subarch_neon=1
    TEST_arch_x86_64_abi=x86_64-little_endian-lp64
    TEST_arch_x86_64_subarch_cx16=1
    TEST_arch_x86_64_subarch_sse3=1
    TEST_arch_x86_64_subarch_sse4.1=1
    TEST_arch_x86_64_subarch_ssse3=1

This is mainly done to provide all ABIs in our module JSON files but
hopefully comes in handy in other situations as well.

Change-Id: Idacd82fc87523657b8bb6ec0c0424f2361572dce
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>