Blacklist failing QNetworkReply test functions on VxWorks
We need to run COIN tests for VxWorks in a blocking mode, but test
functions of QNetworkReply:
 - chaining
 - getFromFile
 - getFromUnreachableIp
 - ioGetFromFile
 - moreActivitySignals
fail on VxWorks, the reason is yet to be examined.

We need to run tests on COIN for VxWorks as significant, and this test
prevents it, to reduce growing number of regressions and allow
iterative introduction of VxWorks as a supported platform.

Blacklist test functions of QNetworkReply failing on VxWorks.

Task-number: QTBUG-115777
Task-number: QTBUG-130740
Change-Id: I3ff5078e2d9073b3f90162a432e9033b0bd46b8f
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>