Silence QResourceEngine::cleanupTestCase failure on VxWorks
VxWorks test run fails cleanupTestCase procedure for not yet known
reason: resource `m_runtimeResourceRcc` exists, but it's failed to be
removed due to refernce count of QDynamicFileResourceRoot instance
equals to 46 (regardless of how many test cases were executed), which
causes `QResource::unregisterResource` to remove false.

Identical situation is with second call to
`QResource::unregisterResource`, but but for the other overload, where
ref counter have some absurd value of 93.

Silent this failure for now on VxWorks by ifdef'ing this problem until
it is fixed for VxWorks, add explicit comment.

Since this is not a test case on itself, blacklisting it can't work, so
remove BLACKLIST for this test as it didn't work nor contain anything
else than `cleanupTestCase` for VxWorks.

Task-number: QTBUG-115777
Task-number: QTBUG-130069
Change-Id: I61fc8fcdd5a3842bcb3ab0d4ea7f3e17c4baaad6
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>