Fix disappearing text with large bitmap fonts
If you add a scale to bitmap fonts, at some point they will hit
the threshold where we fall back to QPainterPath rendering to
save memory. However, this is not supported for bitmap fonts
since there are no paths to draw. The result is that the text
will disappear.

So for non-scalable fonts where the font engine supports applying
the transform, we return true from shouldDrawCachedGlyphs()
instead of checking if the resulting font size is over 64 px as
we do for scalable fonts.

[ChangeLog][Text] Fixed an issue where text in bitmap fonts
would disappear when scaled up to very large sizes.

Fixes: QTBUG-130977
Change-Id: Idfd5517844e5f8e24245e3e09fa86e79783907a1
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>