macOS: Always propagate QCocoaWindow::setGeometry to NSView/NSWindow
We had an optimization that tried to avoid calling setFrame on the
view if the geometry had already been applied, but this logic could
fail if the NSView or NSWindow for various reasons did not reflect
our view of the world through QPlatformWindow::geometry().

One such case was when adding a NSToolBar to a window, in which
case we would end up with the wrong window size.

We now always propagate the geometry, under the assumption that
AppKit will handle the case of setting the same geometry as a noop.

We also pick up any geometry changes that happen during the initial
initializion of the window, even if we are not sending geometry change
events for those until the window is made visible. This ensures we
have an as accurate view of the native window state as possible.

Pick-to: 6.8
Change-Id: I76c753b6f4be8fdbb7a5f6ca3a8b91e013c0b36b
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>