Blacklist tst_QFont::familyNameWithCommaQuote data tags on VxWorks
tst_QFont::familyNameWithCommaQuote test fails for in line 789 on
VxWorks, when test runs with "weird", "commaSeparated" and
"commaSeparatedBogus" data tags. A call to QFontInfo::family returns
"weird" font name instead of standard font name as test expects.

The exact reason is to be found.

We need to run tests on COIN for VxWorks as significant, and this test
prevents it, to reduce growing number of regressions and allow
iterative introduction of VxWorks as a supported platform.

Blacklist failing data tags of familyNameWithCommaQuote tests on VxWorks.

Task-number: QTBUG-115777
Task-number: QTBUG-130738
Change-Id: Ie85c89392d48495226209b5aa76632dfb4c7bc25
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>